private/dbutil/pgutil: move type helpers to separate file
